Implement support colored console output (For Windows and Unix). Disable by default. See comments in mangosd.conf Please test. Fixed arguments types in MapManager::ExistMAP
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1997 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/MapManager.h
@@ -56,7 +56,7 @@ class MANGOS_DLL_DECL MapManager : public MaNGOS::Singleton<MapManager, MaNGOS::
             }
         }
 
-        static bool ExistMAP(int mapid,int x,int y);
+        static bool ExistMAP(int mapid, float x, float y);
     private:
         MapManager();
         ~MapManager();
